requirements.txt
asgiref==3.3.1
Django==3.0
pytz==2021.1
sqlparse==0.4.1
psycopg2-binary>=2.8
mysqlclient>=2.0
django-ckeditor==6.1.0
Pillow
django-filter
django-sortedm2m==3.1.1



